Create testing exclusions, create functions file
