+++ .coveragerc
@@ -1,0 +1,3 @@
+[report]
+exclude_lines =
+    pragma: no cover
