+++ .gitignore
@@ -1,2 +1,3 @@
 __pycache__/
 .coverage
+main.ipynb
